backend/requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dnspython==1.16.0
firebase-admin==3.2.1
Flask==1.1.1
Flask-PyMongo==2.3.0
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.1
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
gunicorn==20.0.4
httplib2==0.17.0
ibm-cloud-sdk-core==1.5.1
ibm-watson==4.2.1
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
msgpack==0.6.2
multidict==4.7.4
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
slackclient==2.5.0
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.8
websocket-client==0.48.0
Werkzeug==0.16.1
yarl==1.4.2
